+++ frontend/src/contexts/FavoritesContext.tsx
@@ -1,22 +1,6 @@
-import React, {
-  createContext,
-  useContext,
-  useState,
-  ReactNode,
-  useEffect,
-} from "react";
+import React, { useState, ReactNode, useEffect } from "react";
 import { Product } from "@/types/product";
-
-interface FavoritesContextType {
-  favorites: Product[];
-  addFavorite: (product: Product) => void;
-  removeFavorite: (productId: string) => void;
-  isFavorite: (productId: string) => boolean;
-}
-
-const FavoritesContext = createContext<FavoritesContextType | undefined>(
-  undefined,
-);
+import { FavoritesContext } from "./favorites";
 
 interface FavoritesProviderProps {
   children: ReactNode;
@@ -64,12 +48,4 @@ export const FavoritesProvider: React.FC<FavoritesProviderProps> = ({
       {children}
     </FavoritesContext.Provider>
   );
-};
-
-export const useFavorites = () => {
-  const context = useContext(FavoritesContext);
-  if (context === undefined) {
-    throw new Error("useFavorites must be used within a FavoritesProvider");
-  }
-  return context;
 };
