+++ frontend/src/contexts/ThemeContext.tsx
@@ -1,13 +1,9 @@
 import React from "react";
 import { useTheme } from "@/hooks/useTheme";
-
-export const ThemeContext = React.createContext({
-  theme: "light",
-  isDark: false,
-});
+import { ThemeContext } from "./theme";
 
 export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
-  const { theme, changeTheme, isDark } = useTheme();
+  const { theme, isDark } = useTheme();
 
   return (
     <ThemeContext.Provider value={{ theme, isDark }}>
