+++ frontend/src/contexts/cart.ts
@@ -1,0 +1,18 @@
+import { createContext } from "react";
+import { Product } from "@/types/product";
+
+interface CartItem extends Product {
+  quantity: number;
+}
+
+interface CartContextType {
+  cartItems: CartItem[];
+  addToCart: (product: Product, quantity?: number) => void;
+  removeFromCart: (productId: string) => void;
+  updateQuantity: (productId: string, quantity: number) => void;
+  clearCart: () => void;
+}
+
+export const CartContext = createContext<CartContextType | undefined>(
+  undefined,
+);
