+++ frontend/src/contexts/favorites.ts
@@ -1,0 +1,13 @@
+import { createContext } from "react";
+import { Product } from "@/types/product";
+
+interface FavoritesContextType {
+  favorites: Product[];
+  addFavorite: (product: Product) => void;
+  removeFavorite: (productId: string) => void;
+  isFavorite: (productId: string) => boolean;
+}
+
+export const FavoritesContext = createContext<FavoritesContextType | undefined>(
+  undefined,
+);
