+++ frontend/src/contexts/theme.ts
@@ -1,0 +1,6 @@
+import React from "react";
+
+export const ThemeContext = React.createContext({
+  theme: "light",
+  isDark: false,
+});
