+++ frontend/src/hooks/useFavorites.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { FavoritesContext } from "@/contexts/favorites";
+
+export const useFavorites = () => {
+  const context = useContext(FavoritesContext);
+  if (context === undefined) {
+    throw new Error("useFavorites must be used within a FavoritesProvider");
+  }
+  return context;
+};
